Update Readme and Add explanation in file InstallPhpcs.php

+++ src/Commands/InstallPhpcs.php
@@ -2,11 +2,12 @@
 
 namespace Botble\GitCommitChecker\Commands;
 
-use RuntimeException;
 use Illuminate\Console\Command;
+use Illuminate\Console\ConfirmableTrait;
 
 class InstallPhpcs extends Command
 {
+    use ConfirmableTrait;
     /**
      * The name and signature of the console command.
      *
@@ -19,7 +20,7 @@ class InstallPhpcs extends Command
      *
      * @var string
      */
-    protected $description = 'Create Phpcs.xml ';
+    protected $description = 'Create default phpcs.xml';
 
     /**
      * Create a new command instance.
@@ -38,28 +39,44 @@ class InstallPhpcs extends Command
      */
     public function handle()
     {
-        $rootphpcs = base_path('phpcs.xml');
         $phpcs = base_path('vendor/botble/git-commit-checker/phpcs.xml');
+        $rootphpcs = base_path('phpcs.xml');
 
+        // Checkout existence of sample phpcs.xml.
+        if (!file_exists($phpcs)) {
+            $this->error('The sample phpcs.xml does not exist! Try to reinstall botble/git-commit-checker package!');
+            return false;
+        }
+
+        // Checkout existence phpcs.xml in root path of project.
         if (file_exists($rootphpcs)) {
             if (!$this->confirmToProceed('phpcs.xml already exists, do you want to overwrite it?', true)) {
                 return false;
-            }            
+            }
+            
+            //remove old phpcs.xml file form root
             unlink($rootphpcs);
-        }   
+        }
 
-        $this->writePHPCS($phpcs, $rootphpcs);
+        $this->writePHPCS($phpcs, $rootphpcs)
+            ? $this->info('Phpcs.xml successfully created!')
+            : $this->error('Unable to create phpcs.xml');
     }
 
+    /**
+     * Copy phpcs.xml file to root and return true on success, false otherwise.
+     *
+     * @param string $phpcs
+     * @param string $rootphpcs
+     * @return bool
+     */
     protected function writePHPCS(string $phpcs, string $rootphpcs): bool
-    {    
-                
-        if (!copy($phpcs, $rootphpcs)) { 
+    {
+        // phpcs.xml file to root
+        if (!copy($phpcs, $rootphpcs)) {
             return false;
-        } 
-        else { 
+        } else {
             return true;
         }
-        
     }
 }
